+++ Algorithms/TrafficAssignment/TravelCostFunctions/InverseFunction.h
@@ -15,13 +15,13 @@ class InverseFunction {
   InverseFunction(const GraphT& graph) : graph(graph) {}
 
   // Returns the travel cost on edge e, given the traffic flow x on e.
-  double operator()(const int e, const double x) const {
+  float operator()(const int e, const float x) const {
     assert(x >= 0);
     return 1.0 * graph.length(e) * (155 / (x + 1) + 0.85) + 0.0 * graph.travelTime(e);
   }
 
   // Returns the derivative of e's travel cost function at x.
-  double derivative(const int e, const double x) const {
+  float derivative(const int e, const float x) const {
     assert(x >= 0);
     return 1.0 * graph.length(e) * -155 / std::pow(x + 1, 2);
   }
